cleaned up DetailScreen
